Fix all build warnings and CTAD error for cross-platform builds
- Add CTAD support for symbol_binding in shared_object.hh
- Add warning suppressions for SDL3 headers (old-style-cast, undef)
- Fix sign conversion warnings in surface.hh, surface.cc, surface_renderer.cc
- Fix implicit int-float conversion warnings in surface_renderer.hh
- Suppress deprecated codecvt_utf8 warnings in hidapi.hh
- Fix unused lambda captures in test files (use constexpr)
- Fix unused variable warning in test_dimension.cc

+++ include/sdlpp/video/surface.hh
@@ -322,9 +322,9 @@ namespace sdlpp {
                         break;
                     case 3:
                         if (SDL_BYTEORDER == SDL_BIG_ENDIAN) {
-                            pixel = p[0] << 16 | p[1] << 8 | p[2];
+                            pixel = static_cast<uint32_t>(p[0]) << 16 | static_cast<uint32_t>(p[1]) << 8 | static_cast<uint32_t>(p[2]);
                         } else {
-                            pixel = p[0] | p[1] << 8 | p[2] << 16;
+                            pixel = static_cast<uint32_t>(p[0]) | static_cast<uint32_t>(p[1]) << 8 | static_cast<uint32_t>(p[2]) << 16;
                         }
                         break;
                     case 4:
